Perf fixes after testing on real repos
